USE IT UNTIL THE PR WILL BE MERGED: https://github.com/STRML/react-grid-layout/pull/1390
----------------------------------------
Fixes #784 #1372 #1376

Horizontal compact mode should not compact vertically by reducing l.y by using the bottom function: It reduces the container and triggers items of the layout with the highest y values to reposition themselves because the repositioning is mixed with the horizontal compaction. So in my opinion, this one line of code needs to be removed (as done in this Pull Request).